+++ producer/src/butter/vector2.h
@@ -23,19 +23,19 @@ namespace Tracer
 {
 	// inline methods
 		// Constructors
-	inline Vector2 vector2(const double _x,const double _y);
-	inline Vector2 vector2(const double _v);
+	inline Vector2 vector2(const float _x,const float _y);
+	inline Vector2 vector2(const float _v);
 
 		// Operators
-	inline Vector2 operator*(const Vector2& _v1, double parFactor);
-	inline Vector2 operator/(const Vector2& _v1, double parFactor);
+	inline Vector2 operator*(const Vector2& _v1, float parFactor);
+	inline Vector2 operator/(const Vector2& _v1, float parFactor);
 	inline Vector2 operator+(const Vector2& _v1, const Vector2& _v2);
 	inline Vector2 operator-(const Vector2& _v1, const Vector2& _v2);
 
 		// Cannonic operations
-	inline double dotProd(const Vector2& _v1, const Vector2& _v2);
-	inline double length(const Vector2& _v1);
-	inline double length2(const Vector2& _v1);
+	inline float dotProd(const Vector2& _v1, const Vector2& _v2);
+	inline float length(const Vector2& _v1);
+	inline float length2(const Vector2& _v1);
 
 	// Default values;
 	extern const Vector2 v2_ZERO;
